requirements.txt
fastapi==0.120.1
uvicorn==0.38.0
python-dotenv==1.2.1
requests==2.32.5
langchain==0.3.27
langchain-core==0.3.79
langchain-community==0.3.30
langchain-google-genai==2.0.10
google-generativeai==0.8.5
streamlit==1.39.0